+++ app/src/main/java/br/ufg/pos/fswmd/myfriendsapp/detailFriend/activity/DetailFriendContract.kt
@@ -16,7 +16,7 @@ interface DetailFriendView {
 
 interface DetailFriendPresenter {
 
-    fun retrieveFriendById(id: Long)
+    fun retrieveFriendById(id: Int)
 
     fun deleteFriend(friend: Friend)
